OPHJOD-1832: Use Chatbot from DS
